rename(Result): domain/result 패키지로 이동
